+++ mocha-requires.js
@@ -2,17 +2,13 @@ import chai from 'chai'
 import chaiAsPromised from 'chai-as-promised'
 import sinonChai from 'sinon-chai'
 import jsonEqual from 'chai-json-equal'
-import {
-  catchPromise,
-  expectThrow,
-  expectNoThrow,
-  expectPromiseNoThrow
-} from './src/test_helpers/utils'
+import { catchPromise, expectThrow, expectNoThrow, expectPromiseNoThrow } from './src/test_helpers/utils'
 chai.use(chaiAsPromised)
 chai.use(sinonChai)
 chai.use(jsonEqual)
 
 global.expect = chai.expect
+
 global.catchPromise = catchPromise
 global.expectThrow = expectThrow
 global.expectNoThrow = expectNoThrow