+++ src/base/operation.js
@@ -19,9 +19,9 @@ import { ManageSaleBuilder } from './operations/manage_sale_builder'
 import { PaymentBuilder } from './operations/payment_builder'
 import { BindExternalSystemAccountIdBuilder } from './operations/bind_external_system_account_id_builder'
 import { CreateManageLimitsRequestBuilder } from './operations/create_manage_limits_request_builder'
-import { CreateAtomicSwapBidCreationRequestBuilder } from './operations/create_atomic_swap_bid_creation_request_builder'
-import { CancelAtomicSwapBidBuilder } from './operations/cancel_atomic_swap_bid_builder'
-import { CreateAtomicSwapRequestBuilder } from './operations/create_atomic_swap_request_builder'
+import { CreateAtomicSwapAskRequestBuilder } from './operations/create_atomic_swap_ask_request_builder'
+import { CancelAtomicSwapAskBuilder } from './operations/cancel_atomic_swap_ask_builder'
+import { CreateAtomicSwapBidRequestBuilder } from './operations/create_atomic_swap_bid_request_builder'
 import { CreateWithdrawRequestBuilder } from './operations/create_withdraw_request_builder'
 import { ManageLimitsBuilder } from './operations/manage_limits_builder'
 import { ManageKeyValueBuilder } from './operations/manage_key_value_builder'
@@ -33,6 +33,9 @@ import { ManagePollBuilder } from './operations/manage_poll_builder'
 import { ManageAccountSpecificRuleBuilder } from './operations/manage_account_specific_rule_builder'
 import { CancelChangeRoleRequestBuilder } from './operations/cancel_change_role_request_builder'
 import { ManageAccountRoleBuilder } from './operations/manage_account_role_builder'
+import { RemoveAssetPairOpBuilder } from './operations/remove_asset_pair_op_builder'
+import { InitiateKYCRecoveryBuilder } from './operations/initiate_kyc_recovery_builder'
+import { CreateKYCRecoveryRequestBuilder } from './operations/create_kyc_recovery_request_builder'
 
 export class Operation extends BaseOperation {
   /**
@@ -430,15 +433,14 @@ export class Operation extends BaseOperation {
       case xdr.OperationType.createManageLimitsRequest():
         CreateManageLimitsRequestBuilder.createManageLimitsRequestToObject(result, attrs)
         break
-      case xdr.OperationType.createAswapBidRequest():
-        CreateAtomicSwapBidCreationRequestBuilder
-          .createASwapBidCreationRequestToObject(result, attrs)
+      case xdr.OperationType.createAtomicSwapBidRequest():
+        CreateAtomicSwapBidRequestBuilder.createAtomicSwapBidRequestToObject(result, attrs)
         break
-      case xdr.OperationType.cancelAswapBid():
-        CancelAtomicSwapBidBuilder.cancelASwapBidToObject(result, attrs)
+      case xdr.OperationType.cancelAtomicSwapAsk():
+        CancelAtomicSwapAskBuilder.cancelAtomicSwapAskToObject(result, attrs)
         break
-      case xdr.OperationType.createAswapRequest():
-        CreateAtomicSwapRequestBuilder.createASwapRequestToObject(result, attrs)
+      case xdr.OperationType.createAtomicSwapAskRequest():
+        CreateAtomicSwapAskRequestBuilder.createAtomicSwapAskRequestToObject(result, attrs)
         break
       case xdr.OperationType.stamp():
         StampBuilder.stampToObject(result, attrs)
@@ -463,6 +465,15 @@ export class Operation extends BaseOperation {
         break
       case xdr.OperationType.manageAccountRole():
         ManageAccountRoleBuilder.manageAccountRoleToObject(result, attrs)
+        break
+      case xdr.OperationType.removeAssetPair():
+        RemoveAssetPairOpBuilder.removeAssetPairOpToObject(result, attrs)
+        break
+      case xdr.OperationType.initiateKycRecovery():
+        InitiateKYCRecoveryBuilder.initiateKYCRecoveryToObject(result, attrs)
+        break
+      case xdr.OperationType.createKycRecoveryRequest():
+        CreateKYCRecoveryRequestBuilder.createKYCRecoveryRequestOpToObject(result, attrs)
         break
       default:
         throw new Error('Unknown operation')