+++ src/base/operations/create_atomic_swap_request_builder.spec.js
@@ -1,22 +1,24 @@
 import { Operation } from '../operation'
-import { CreateAtomicSwapRequestBuilder } from './create_atomic_swap_request_builder'
+import { CreateAtomicSwapBidRequestBuilder } from './create_atomic_swap_bid_request_builder'
 import { default as xdr } from '../generated/xdr_generated'
+import { isEqual } from 'lodash'
 
-describe('Create ASwapBidCreation request', function () {
+describe('Create ASwapBid request', function () {
   it('Success', function () {
     let opts = {
       baseAmount: '911',
-      bidID: '69',
+      askID: '69',
       quoteAsset: 'ETH',
-      creatorDetails: ''
+      creatorDetails: { data: 'new atomic swap' }
     }
-    let op = CreateAtomicSwapRequestBuilder.createASwapRequest(opts)
+    let op = CreateAtomicSwapBidRequestBuilder.createAtomicSwapBidRequest(opts)
     let xdrOp = op.toXDR('hex')
     let operation = xdr.Operation.fromXDR(Buffer.from(xdrOp, 'hex'))
     let obj = Operation.operationToObject(operation)
-    expect(obj.type).to.be.equal(xdr.OperationType.createAswapRequest().name)
+    expect(obj.type).to.be.equal(xdr.OperationType.createAtomicSwapBidRequest().name)
     expect(obj.baseAmount).to.be.equal(opts.baseAmount)
-    expect(obj.bidID).to.be.equal(opts.bidID)
+    expect(obj.askID).to.be.equal(opts.askID)
     expect(obj.quoteAsset).to.be.equal(opts.quoteAsset)
+    expect(isEqual(opts.creatorDetails, obj.creatorDetails)).to.be.true
   })
 })