+++ src/const/key-value.const.js
@@ -5,7 +5,12 @@ export const KEY_VALUE_KEYS = Object.freeze({
   saleCreateTasks: 'sale_create_tasks:*',
   withdrawalTasks: 'withdrawal_tasks',
   amlAlertTasks: 'aml_alert_tasks',
+  atomicSwapBidTasks: 'atomic_swap_bid_tasks:*',
+  atomicSwapAskTasks: 'atomic_swap_ask_tasks',
   paymentCreateTasks: 'payment_create_tasks:*',
   createPollTasks: 'create_poll_tasks:*',
-  change_role_tasks: 'change_role_tasks:*:*'
+  change_role_tasks: 'change_role_tasks:*:*',
+  kycRecoveryEnabled: 'kyc_recovery_enabled',
+  kycRecoverySignerRole: 'kyc_recovery_signer_role',
+  createKycRecoveryTasks: 'create_kyc_recovery_tasks'
 })