+++ test/helpers/index.js
@@ -17,10 +17,14 @@ import { Withdraw } from './withdraw'
 import { Payment } from './payment'
 import { StampHelper } from './stamp'
 import { LicenseHelper } from './license'
+import { AtomicSwapAsk } from './atomic_swap_ask'
+import { AtomicSwapBid } from './atomic_swap_bid'
 import { Poll } from './poll'
 import { Vote } from './vote'
 import { ManageAccountRole } from './manage_account_role'
 import { ChangeRole } from './change_role'
+import { KYCRecoveryHelper } from './kyc_recovery'
+
 
 export const sdk = new TokenD(config.api_url, {
   allowHttp: config.allow_http
@@ -51,7 +55,10 @@ export const withdrawHelper = new Withdraw(DEFAULTS)
 export const amlAlertHelper = new AmlAlert(DEFAULTS)
 export const stampHelper = new StampHelper(DEFAULTS)
 export const licenseHelper = new LicenseHelper(DEFAULTS)
+export const atomicSwapAskHelper = new AtomicSwapAsk(DEFAULTS)
+export const atomicSwapBidHelper = new AtomicSwapBid(DEFAULTS)
 export const pollHelper = new Poll(DEFAULTS)
 export const voteHelper = new Vote(DEFAULTS)
 export const manageAccountRoleHelper = new ManageAccountRole(DEFAULTS)
 export const changeRoleHelper = new ChangeRole(DEFAULTS)
+export const kycRecoveryHelper = new KYCRecoveryHelper(DEFAULTS)