+++ CHANGELOG.md
@@ -9,6 +9,32 @@ for further information about branching and tagging conventions.
 
 ## [Unreleased]
 ### Added
+- integration test for atomic swap
+- start using utils/validators
+- Operation to remove asset pair
+- Builders for `InitiateKYCRecovery`, `CreateKYCRecoveryRequest`
+- Unit tests for for `InitiateKYCRecovery`, `CreateKYCRecoveryRequest`
+- Integration tests for `CreateKYCRecoveryRequest`, `CreateKYCRecoveryRequest`
+
+### Changed
+- xdr, version 3.4.0-x.0
+
+### Fixed
+- operation builders related to atomic swap
+
+## [1.7.0-rc.2] - 2019-05-30
+### Fixed
+- Transaction error parsing if no messages are provided
+- Bug with deprecated `requestDetails` field of
+  `ReviewRequestBuilder#reviewWithdrawRequest`. Now this field be set
+  automatically under the hood
+
+## [1.7.0-rc.1] - 2019-05-21
+### Changed
+- Now an empty object on creator details validator is allowed
+
+## [1.7.0-rc.0] - 2019-05-21
+### Added
 - Validators for specified types:
   - Undefined
   - Array
@@ -21,7 +47,7 @@ for further information about branching and tagging conventions.
   - XDR enum type
   - Creator details
 - Check for spaces in `BaseOperation.isValidString`
-- `DocumentsManager` class to encapsulate creating document's config & 
+- `DocumentsManager` class to encapsulate creating document's config &
   uploading a file to the storage
 
 ### Changed
@@ -140,7 +166,10 @@ for further information about branching and tagging conventions.
 
 ## [1.3.0] - 2019-03-01
 
-[Unreleased]: https://github.com/tokend/new-js-sdk/compare/1.7.0-x.2...HEAD
+[Unreleased]: https://github.com/tokend/new-js-sdk/compare/1.7.0-rc.2...HEAD
+[1.7.0-rc.2]: https://github.com/tokend/new-js-sdk/compare/1.7.0-rc.1...1.7.0-rc.2
+[1.7.0-rc.1]: https://github.com/tokend/new-js-sdk/compare/1.7.0-rc.0...1.7.0-rc.1
+[1.7.0-rc.0]: https://github.com/tokend/new-js-sdk/compare/1.7.0-x.2...1.7.0-rc.0
 [1.7.0-x.2]: https://github.com/tokend/new-js-sdk/compare/1.7.0-x.1...1.7.0-x.2
 [1.7.0-x.1]: https://github.com/tokend/new-js-sdk/compare/1.7.0-x.0...1.7.0-x.1
 [1.7.0-x.0]: https://github.com/tokend/new-js-sdk/compare/1.6.0...1.7.0-x.0